[FEATURE] Teapot and Teacup Load Balancer.
- Remove conflicting packages with licence.
- Create Load Balancer class.
