+++ libs/constants.ts
@@ -44,9 +44,7 @@ export enum BOOT_STAGES {
 export const BOOT_ORDER = [
   BOOT_STAGES.BOOT_DEPENDENCIES,
   BOOT_STAGES.INITIALIZE_MIDDLEWARES,
-  // BOOT_STAGES.APPLICATION,
-  // BOOT_STAGES.AFTER_APPLICATION_MIDDLEWARES,
-  // BOOT_STAGES.START
+  BOOT_STAGES.APPLICATION
 ];
 
 export const BOOT_STAGES_LIST = [
